Dear Boost,
We think we need to spend some time to think about where this relationship is going. We've grown apart of over the years and we are not quite sure that this was meant to be.

While we're separated, feel free to see other people, other groups, different cryptocurrencies. We'll certainly play with other libraries that will very likely replace you in the end. There's still a chance though that we will come running back to you so definitely keep your hopes up. If you happen to find someone else that loves you as much as you need loved, go, be free.

For now, you can keep the boxes in the closet but at some point you're going to need to come and get them.

Sincerely,

The TurtleCoin Developers